IndexedPriorityQueue 구현, 탐색 싸이클 구현

+++ ServerLib/Types.h
@@ -20,7 +20,7 @@
 
 using SessionIdType = BYTE;
 using PacketSizeType = BYTE;
-using NetworkObjectIdType = size_t;
+using NetworkObjectIdType = unsigned short;
 
 using ExtraInfo = std::variant<void*, SOCKET, unsigned long long, HANDLE>;
